feat: moving bitrate func to avfoundationplayback file
